+++ springboot-backend/src/main/java/com/group/travels/api/travel/TravelResponse.java
@@ -10,7 +10,7 @@ public record TravelResponse(Long id,
                              LocalDateTime travelEndDate,
                              Integer maxNumberOfPlaces,
                              Integer numberOfFreePlaces,
-                             Integer price,
+                             Double price,
                              Long countryID) {
 
     public TravelResponse(Travel travel){
